pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "snapsort"
version = "0.1.0"
description = "Organize images by duplicates and blurriness"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "snapSort" }]
dependencies = [
  "Pillow>=10.0.0",
  "imagehash>=4.3.1",
  "opencv-python>=4.9.0.0",
  "numpy>=1.26.0",
  "tqdm>=4.66.0",
  "rich>=13.7.0",
  # Optional: camera RAW support (.nef) via LibRaw
  "rawpy>=0.19.0",
]

[project.scripts]
snapsort = "snapsort.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
